fraud_detection_system/requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Validation
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0.0
python-box>=7.0.0
ensure>=1.0.0

# Model Training
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
optuna>=3.0.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0

# MLOps
mlflow>=2.8.0
evidently>=0.4.0

# Database
pymongo>=4.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
joblib>=1.3.0

# Scientific / Cloud
scipy>=1.11.0
boto3>=1.28.0
tqdm>=4.66.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
colorlog>=6.7.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0


# Existing dependencies...

# Imbalanced Learning
imbalanced-learn>=0.11.0

# Feature Selection
scikit-learn>=1.3.0

# Scipy for statistical operations
scipy>=1.11.0


# Existing dependencies...

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting
slowapi>=0.1.8

# Templates
jinja2>=3.1.2
aiofiles>=23.2.1

# Kafka
confluent-kafka>=2.3.0


# Monitoring
evidently>=0.4.0
mlflow>=2.8.0

# Dashboard
dash>=2.14.0
plotly>=5.18.0

# Optional: Additional monitoring
prometheus-client>=0.19.0